+++ src/components/Login.vue
@@ -1,30 +1,30 @@
 <template>
   <div
-    class="relative flex flex-column align-items-center justify-content-center h-full m-2"
+    class="auth-screen relative flex flex-column align-items-center justify-content-center h-full m-2"
   >
-    <Panel v-if="newWalletForm" class="col-12 md:col-8 lg:col-6">
+    <Panel v-if="newWalletForm" class="auth-panel">
       <template #header>
         {{ $t("login.new_wallet") }}
       </template>
       <form autocomplete="off" @submit="createWalletClick">
         <div class="field grid">
-          <label for="newwallet-name" class="col-12 mb-2 md:col-2 md:mb-0">
+          <label for="newwallet-name" class="col-12 mb-1">
             {{ $t("login.new_wallet_name") }}
           </label>
-          <div class="col-12 md:col-10">
+          <div class="col-12">
             <InputText
               id="newwallet-name"
               v-model="newname"
-              class="w-full my-2"
+              class="w-full"
               autocomplete="off"
             />
           </div>
         </div>
         <div class="field grid">
-          <label for="newwallet-pass" class="col-12 mb-2 md:col-2 md:mb-0">
+          <label for="newwallet-pass" class="col-12 mb-1">
             {{ $t("login.new_wallet_pass") }}
           </label>
-          <div class="col-12 md:col-10">
+          <div class="col-12">
             <Password
               inputId="newwallet-pass"
               v-model="pass"
@@ -37,51 +37,56 @@
           </div>
         </div>
         <div class="field grid">
-          <div class="col-12 mb-2 md:col-2 md:mb-0"></div>
-          <div class="col-12 md:col-10">
-            <Button type="submit" id="new_wallet_button_create">
+          <div class="col-12">
+            <Button type="submit" id="new_wallet_button_create" class="w-full">
               {{ $t("login.new_wallet_button_create") }}
             </Button>
-            <router-link to="/import-wallet" class="mx-2">
-              <Button severity="secondary" id="new_wallet_button_import">
-                {{ $t("login.new_wallet_button_import") }}
+            <div class="flex gap-2 mt-2">
+              <router-link to="/import-wallet" class="flex-1">
+                <Button
+                  severity="secondary"
+                  id="new_wallet_button_import"
+                  class="w-full"
+                >
+                  {{ $t("login.new_wallet_button_import") }}
+                </Button>
+              </router-link>
+              <Button
+                v-if="wallets.length > 0"
+                severity="secondary"
+                class="flex-1"
+                @click="newWalletForm = false"
+                id="go_back"
+              >
+                {{ $t("global.go_back") }}
               </Button>
-            </router-link>
-            <Button
-              v-if="wallets.length > 0"
-              severity="secondary"
-              class="mx-2"
-              @click="newWalletForm = false"
-              id="go_back"
-            >
-              {{ $t("global.go_back") }}
-            </Button>
-            <p class="my-2">
+            </div>
+            <p class="my-2 auth-help">
               {{ $t("login.new_wallet_help1") }}
             </p>
-            <p class="my-2">
+            <p class="my-2 auth-help">
               {{ $t("login.new_wallet_help2") }}
             </p>
           </div>
         </div>
       </form>
     </Panel>
-    <Panel v-if="!newWalletForm" class="col-12 md:col-8 lg:col-6">
+    <Panel v-if="!newWalletForm" class="auth-panel">
       <template #header>
         {{ $t("login.open_wallet") }}
       </template>
       <form @submit="auth">
         <div class="field grid">
-          <label for="wallet-select" class="col-12 mb-2 md:col-2 md:mb-0">{{
+          <label for="wallet-select" class="col-12 mb-1">{{
             $t("login.select_wallet")
           }}</label>
-          <div class="col-12 md:col-10">
+          <div class="col-12">
             <Select
               id="wallet-select"
               v-model="wallet"
               :options="wallets"
               filter
-              class="my-2 w-full"
+              class="w-full"
               optionLabel="name"
               optionValue="code"
               :placeholder="$t('login.select_wallet')"
@@ -89,14 +94,14 @@
           </div>
         </div>
         <div class="field grid">
-          <label for="wallet-pass" class="col-12 mb-2 md:col-2 md:mb-0">{{
+          <label for="wallet-pass" class="col-12 mb-1">{{
             $t("login.wallet_password")
           }}</label>
-          <div class="col-12 md:col-10">
+          <div class="col-12">
             <Password
               inputId="wallet-pass"
               v-model="pass"
-              class="w-full my-2"
+              class="w-full"
               inputClass="w-full"
               :feedback="false"
               autocomplete="new-password"
@@ -104,9 +109,8 @@
           </div>
         </div>
         <div class="field grid">
-          <label for="wallet-pass" class="col-12 mb-2 md:col-2 md:mb-0"></label>
-          <div class="col-12 md:col-10">
-            <Button type="submit" id="new_wallet_button_open">
+          <div class="col-12">
+            <Button type="submit" id="new_wallet_button_open" class="w-full">
               {{ $t("login.new_wallet_button_open") }}
             </Button>
           </div>