+++ src/components/NewWallet.vue
@@ -1,30 +1,30 @@
 <template>
   <div
-    class="relative flex flex-column align-items-center justify-content-center h-full m-2"
+    class="auth-screen relative flex flex-column align-items-center justify-content-center h-full m-2"
   >
-    <Panel class="col-12 md:col-8 lg:col-6">
+    <Panel class="auth-panel">
       <template #header>
         {{ $t("login.new_wallet") }}
       </template>
       <form autocomplete="off" @submit="createWalletClick">
         <div class="field grid">
-          <label for="newwallet-name" class="col-12 mb-2 md:col-2 md:mb-0">
+          <label for="newwallet-name" class="col-12 mb-1">
             {{ $t("login.new_wallet_name") }}
           </label>
-          <div class="col-12 md:col-10">
+          <div class="col-12">
             <InputText
               id="newwallet-name"
               v-model="newname"
-              class="w-full my-2"
+              class="w-full"
               autocomplete="off"
             />
           </div>
         </div>
         <div class="field grid">
-          <label for="newwallet-pass" class="col-12 mb-2 md:col-2 md:mb-0">
+          <label for="newwallet-pass" class="col-12 mb-1">
             {{ $t("login.new_wallet_pass") }}
           </label>
-          <div class="col-12 md:col-10">
+          <div class="col-12">
             <Password
               inputId="newwallet-pass"
               v-model="pass"
@@ -37,15 +37,14 @@
           </div>
         </div>
         <div class="field grid">
-          <div class="col-12 mb-2 md:col-2 md:mb-0"></div>
-          <div class="col-12 md:col-10">
-            <Button type="submit" id="new_wallet_button_create">
+          <div class="col-12">
+            <Button type="submit" id="new_wallet_button_create" class="w-full">
               {{ $t("login.new_wallet_button_create") }}
             </Button>
-            <p class="my-2">
+            <p class="my-2 auth-help">
               {{ $t("login.new_wallet_help1") }}
             </p>
-            <p class="my-2">
+            <p class="my-2 auth-help">
               {{ $t("login.new_wallet_help2") }}
             </p>
           </div>